Enable compiler on NetBSD, DragonFly, illumos and Solaris
