+++ README.md
@@ -108,7 +108,8 @@ We also test cross compilation for many `GOOS` and `GOARCH` combinations.
   * macOS is tested only on arm64.
 * Compiler
   * Linux is tested on amd64 (native) as well arm64 via emulation.
-  * Windows and FreeBSD are only tested on amd64.
+  * Windows, FreeBSD, NetBSD, DragonFly BSD, illumos and Solaris are
+    tested only on amd64.
   * macOS is tested only on arm64.
 
 wazero has no dependencies and doesn't require CGO. This means it can also be
