+++ config_unsupported.go
@@ -1,5 +1,5 @@
 // This is the opposite constraint of config_supported.go
-//go:build !(amd64 || arm64) || !(darwin || linux || freebsd || windows)
+//go:build !(amd64 || arm64) || !(linux || darwin || freebsd || netbsd || dragonfly || windows)
 
 package wazero
